segregate variables for mobile api
